+++ src/fields/EntriesSearch.php
@@ -35,9 +35,9 @@ class EntriesSearch extends Entries
     /**
      * @inheritdoc
      */
-    protected function inputHtml(mixed $value, ?ElementInterface $element = null): string
+    protected function inputHtml(mixed $value, ?ElementInterface $element = null, bool $inline = false): string
     {
         Craft::$app->getView()->registerAssetBundle(ToolsAsset::class);
-        return parent::inputHtml($value, $element);
+        return parent::inputHtml($value, $element, $inline);
     }
 }